requirements.txt
Django
djangorestframework
djangorestframework_simplejwt
django_ckeditor_5
django_filter
drf-spectacular
pillow
django_environ
markdown
django_debug_toolbar